requirements.txt

pdfminer3k==1.3.1
PyExecJS==1.5.0
python-docx==0.8.5
requests==2.18.4
